% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_video.R
\name{download_video}
\alias{download_video}
\title{Download Video Using youtube-dl}
\usage{
download_video(video_url, prefix = NULL)
}
\arguments{
\item{video_url}{A character. URL for the video to download.}

\item{prefix}{A character. Text to prefix the video title.}
}
\description{
Make sure youtube-dl is installed and set in the path variables!
}
