% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_highlights.R
\name{get_hightlights}
\alias{get_hightlights}
\title{Extract A User's Highlights}
\usage{
get_hightlights(user_id = Sys.getenv("TWITCH_USER_ID"))
}
\arguments{
\item{user_id}{A numeric. Twitch User ID to pull}
}
\value{
A tibble
}
\description{
Extract A User's Highlights
}
