% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitch_auth.R
\name{twitch_auth}
\alias{twitch_auth}
\title{Get Auth Token}
\usage{
twitch_auth(
  client_id = Sys.getenv("TWITCH_CLIENT_ID"),
  client_secret = Sys.getenv("TWITCH_SECRET")
)
}
\arguments{
\item{client_id}{A character. Twitch App ID.}

\item{client_secret}{A character. Twitch App Secret.}
}
\value{
Invisible authorization data.
}
\description{
Thanks to https://github.com/Freguglia/rTwitchAPI/blob/master/R/Utilities.R for getting me started!
}
